feat - move Currency class to money package
