+++ src/main/java/co/invest72/exchange_rate/domain/service/ExchangeRateService.java
@@ -6,7 +6,7 @@ import java.util.Map;
 import java.util.Optional;
 import java.util.concurrent.ConcurrentHashMap;
 
-import co.invest72.exchange_rate.domain.Currency;
+import co.invest72.money.domain.Currency;
 import co.invest72.money.domain.CurrencyPair;
 
 public class ExchangeRateService {
