+++ src/main/java/co/invest72/money/domain/CurrencyPair.java
@@ -2,7 +2,6 @@ package co.invest72.money.domain;
 
 import java.util.Objects;
 
-import co.invest72.exchange_rate.domain.Currency;
 import lombok.EqualsAndHashCode;
 
 @EqualsAndHashCode
