+++ src/test/java/co/invest72/exchange_rate/domain/service/ExchangeRateServiceTest.java
@@ -7,7 +7,7 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import co.invest72.exchange_rate.domain.Currency;
+import co.invest72.money.domain.Currency;
 import co.invest72.money.domain.CurrencyPair;
 
 class ExchangeRateServiceTest {
